DateCalculator.java
25.	In a Management College a project was allotted. The project should be submitted on a fixed date after a given number of days. WAP to input project allotment date in the format Date: DD/MM/YYYY and number of days i.e. day when the project is to be submitted. 
Print project submission date in the format DD/MM/YYYY
Eg: 	Input: 
Project allotment date: 20/10/2007
	Number of days: 60
	Output: 
	Project submission date: 18/12/2007



or


Design a program to accept a day number (between 1 and 366), year (in 4 digits) from the user to generate and display the corresponding date. Also, accept 'N' (1 <= N <= 100) from the user to compute and display the future date corresponding to 'N' days after the generated date. Display an error message if the value of the day number, year and N are not within the limit or not according to the condition specified.